refactor(FR #98): centralize Ireland data to variants directory (#102)
Move Ireland-specific data files from src/data/ to
src/config/variants/ireland/data/ for better maintainability:

File moves:
- data-centers-ireland.ts → data/data-centers.ts
- tech-hqs-ireland.ts → data/tech-hqs.ts
- irish-unicorns.ts → data/unicorns.ts
- semiconductor-hubs.ts → data/semiconductor-hubs.ts

Also:
- Create data/index.ts for unified exports
- Update all import paths in DeckGLMap.ts
- Update all test file imports

Closes #98
