+++ src/config/variants/ireland/data/data-centers.ts
