+++ src/config/variants/ireland/data/index.ts
@@ -1,0 +1,30 @@
+/**
+ * Ireland Data Exports
+ *
+ * Centralized exports for all Ireland-specific data.
+ * Use this index for cleaner imports across the codebase.
+ */
+
+// Data Centers
+export {
+  IRELAND_DATA_CENTERS,
+  type IrelandDataCenter,
+} from './data-centers';
+
+// Tech HQs (EMEA)
+export {
+  IRELAND_TECH_HQS,
+  type IrelandTechHQ,
+} from './tech-hqs';
+
+// Irish Unicorns
+export {
+  IRISH_UNICORNS,
+  type IrishUnicorn,
+} from './unicorns';
+
+// Semiconductor Hubs
+export {
+  IRELAND_SEMICONDUCTOR_HUBS,
+  type SemiconductorHub,
+} from './semiconductor-hubs';
