+++ src/config/variants/ireland/data/semiconductor-hubs.ts
