+++ src/config/variants/ireland/data/tech-hqs.ts
